Module Lrt.Matcher

Pattern matching on runtime types.

module type S = sig ... end

A matcher with a given return type.

module Make : functor (Return : sig ... end) -> S with type 'a return = 'a Return.t

Instantiate a matcher with a result type.

Example

We will match on these example types.

type t0 = string list

and 'a t1 = 'a array

and ('a, 'b) t2 = ('a * 'b) option [@@deriving t]

The example pattern match should print the type. An appropriate result type is unit -> unit.

module Matcher = Matcher.Make (struct type 'a t = unit -> unit end)

The different cases are registered one by one. Free variables will be substituted in the returned result.

let m =
  let open Matcher in
  let pp_ty = Ttype.print
  empty ~modulo_props:true
  |> add ~t:[%t: string list] (fun () ->
         Format.printf "t0 = %a\n%!" pp_ty t0_t )
  |> add1
       ( module struct
         type 'a t = 'a t1 [@@deriving t]

         let return a_t () =
           Format.printf "%a t1 = %a\n%!" pp_ty a_t pp_ty (t1_t a_t)
       end )
  |> add2
       ( module struct
         type ('a, 'b) t = ('a, 'b) t2 [@@deriving t]

         let return a_t b_t () =
           Format.printf "(%a, %a) t2 = %a\n%!" pp_ty a_t pp_ty b_t pp_ty
             (t2_t a_t b_t)
       end )

The handling of matcher results needs some boilerplate code.

let apply : type a. Matcher.t -> t:a Ttype.t -> unit =
 fun matcher ~t ->
  let open Matcher in
  match apply matcher ~t with
  | None -> print_endline "Not found"
  | Some (M0 (module M : M0 with type matched = a)) -> M.return ()
  | Some (M1 (module M : M1 with type matched = a)) -> M.return ()
  | Some (M2 (module M : M2 with type matched = a)) -> M.return ()

Now everything is set up and the matcher is ready for application.

let () =
  apply m ~t:[%t: t0] ;
  apply m ~t:[%t: int t1] ;
  apply m ~t:[%t: bool t1] ;
  apply m ~t:[%t: float option] ;
  apply m ~t:[%t: (float, string) t2] ;
  apply m ~t:[%t: (unit, string) t2]

The above example program produces the following output.

t0 = string list
int t1 = int array
bool t1 = bool array
Not found
(float, string) t2 = (float * string) option
(unit, string) t2 = (unit * string) option