Module type S.M1

A matching result with one substituted variable.

include Lrt.Unify.T1

type 'a t

val t : 'a Lrt.Ttype.t -> 'a t Lrt.Lrt_ppx_runtime.Types.Ttype.t

type a

First type variable.

type matched

The type given to apply.

val eq : (a t, matched) Lrt.TypEq.t

Proofs the equality of the added type and the matched type to the type checker.

val return : a t return

The result of the match with substituted type variable. Corresponds to the right hand side of -> in usual OCaml pattern matching.