Module Lrt__Json

Json compatible representation of values.

Representation of JSON trees

type number =

| I of int

| F of float

type value =

| Null

| Bool of bool

| Number of number

| String of string

| Array of value list

| Object of (string * value) list

val value_t : value Lrt.Lrt_ppx_runtime.Types.Ttype.t

Conversion between JSON trees and OCaml values

type 'a conv = {

to_json : 'a -> value;

of_json : value -> 'a;

}

Custom mapping for specific types

Custom conv pairs can be globally registered using the add function and its variations. By default, conv t uses the globally registered converters where possible. The default can be overwritten, by providing a custom matcher locally (via ctx).

Note:

It is not allowed to use null in the JSON representation of values, at least if the type is used under the option type constructor (null is reserved for representing the None case).

module Matcher : Lrt.Matcher.S with type 'a return := 'a conv

val add : t:'a Lrt.Ttype.t -> 'a conv -> unit

See Matcher.add. Modifies the global registry.

val add0 : (module Matcher.C0) -> unit

See Matcher.add0. Modifies the global registry.

val add1 : (module Matcher.C1) -> unit

See Matcher.add1. Modifies the global registry.

val add2 : (module Matcher.C2) -> unit

See Matcher.add2. Modifies the global registry.

val matcher : unit -> Matcher.t

Fetch the current global registry of converters.

type ctx

val ctx : ?⁠to_json_field:(string -> string) -> ?⁠matcher:Matcher.t -> unit -> ctx

The default mapping can be manipulated by providing a context to the conv function. The following variations are currently possible.

to_json_field: how to translate a record field name to a JSON field name.

matcher: instead of the global registry of custom converters, use modified one. Allows to overwrite global converters locally.

val conv : ?⁠ctx:ctx -> 'a Lrt.Ttype.t -> 'a conv

Generates a conv pair from the provided type.

Mapping between JSON trees and their textual representation.

val encode : ?⁠utf8:bool -> value -> string

Encode JSON tree into a compact JSON text (single line, etc).

val decode : ?⁠utf8:bool -> ?⁠filename:string -> string -> value

Parse a JSON text into JSON tree. Report parse errors using the Failure exception. The optional filename argument is used to report locations in error messages.

val to_pretty_string : value -> string

Prints a JSON tree into human-friendly text (multi-line, indentation). The output is NOT a parsable json string.

Miscellaneous

module Access : sig ... end

Access values in JSON trees.